Refactor project structure: Split utils into modular packages
